+++ api/app/domain/compute_engines/compute_engine.py
@@ -1,31 +1,17 @@
-from enum import Enum
 from functools import partial
 from typing import Any, List, Literal
 
 import duckdb
 import numpy as np
 import pandas as pd
-import xarray as xr
+from app.domain.models.dataset import Dataset
+from app.domain.repositories.data_api_aoi_geometry_repository import (
+    DataApiAoiGeometryRepository,
+)
+from app.domain.repositories.zarr_dataset_repository import ZarrDatasetRepository
 from app.models.common.areas_of_interest import AreaOfInterest
 from app.models.common.base import StrictBaseModel
 from flox.xarray import xarray_reduce
-from shapely import Geometry
-from shapely.geometry import mapping
-
-
-class Dataset(Enum):
-    tree_cover_loss = "tree_cover_loss"
-    canopy_cover = "canopy_cover"
-    area_hectares = "area_hectares"
-
-    def get_field_name(self):
-        DATASET_TO_NAMES = {
-            Dataset.area_hectares: "area_ha",
-            Dataset.tree_cover_loss: "loss_year",
-            Dataset.canopy_cover: "canopy_cover",
-        }
-
-        return DATASET_TO_NAMES[self]
 
 
 class DatasetFilter(StrictBaseModel):
@@ -83,43 +69,6 @@ class PrecalcHandler:
         sql = f"SELECT id, {groupby_fields}, {agg} FROM data_source WHERE {filters} GROUP BY id, {groupby_fields}"
 
         return await self.precalc_query_service.execute(data_source, sql)
-
-
-class ZarrDatasetRepository:
-    ZARR_LOCATIONS = {
-        Dataset.area_hectares: "s3://gfw-data-lake/umd_area_2013/v1.10/raster/epsg-4326/zarr/pixel_area.zarr",
-        Dataset.tree_cover_loss: "s3://gfw-data-lake/umd_tree_cover_loss/v1.12/raster/epsg-4326/zarr/year.zarr",
-        Dataset.canopy_cover: "s3://gfw-data-lake/umd_tree_cover_density_2000/v1.8/raster/epsg-4326/zarr/threshold.zarr",
-    }
-
-    def load(self, dataset: Dataset, geometry: Geometry = None) -> xr.DataArray:
-        xarr = xr.open_zarr(
-            self.ZARR_LOCATIONS[dataset],
-            storage_options={"requester_pays": True},
-        ).band_data
-        xarr.rio.write_crs("EPSG:4326", inplace=True)
-        xarr.name = dataset.get_field_name()
-
-        if geometry is not None:
-            return self._clip_xarr_to_geometry(xarr, geometry)
-        return xarr
-
-    def _clip_xarr_to_geometry(self, xarr, geom):
-        sliced = xarr.sel(
-            x=slice(geom.bounds[0], geom.bounds[2]),
-            y=slice(geom.bounds[3], geom.bounds[1]),
-        )
-        if "band" in sliced.dims:
-            sliced = sliced.squeeze("band")
-
-        geojson = mapping(geom)
-        clipped = sliced.rio.clip([geojson])
-        return clipped
-
-
-class DataApiAoiGeometryRepository:
-    def load(self, aoi_type: str, aoi_ids: List[str]) -> Geometry:
-        pass
 
 
 class FloxOTFHandler:
