+++ api/app/domain/repositories/data_api_aoi_geometry_repository.py
@@ -1,0 +1,8 @@
+from typing import List
+
+from shapely import Geometry
+
+
+class DataApiAoiGeometryRepository:
+    def load(self, aoi_type: str, aoi_ids: List[str]) -> Geometry:
+        pass
