+++ api/app/domain/repositories/zarr_dataset_repository.py
@@ -1,0 +1,36 @@
+import xarray as xr
+from app.domain.models.dataset import Dataset
+from shapely import Geometry
+from shapely.geometry import mapping
+
+
+class ZarrDatasetRepository:
+    ZARR_LOCATIONS = {
+        Dataset.area_hectares: "s3://gfw-data-lake/umd_area_2013/v1.10/raster/epsg-4326/zarr/pixel_area.zarr",
+        Dataset.tree_cover_loss: "s3://gfw-data-lake/umd_tree_cover_loss/v1.12/raster/epsg-4326/zarr/year.zarr",
+        Dataset.canopy_cover: "s3://gfw-data-lake/umd_tree_cover_density_2000/v1.8/raster/epsg-4326/zarr/threshold.zarr",
+    }
+
+    def load(self, dataset: Dataset, geometry: Geometry = None) -> xr.DataArray:
+        xarr = xr.open_zarr(
+            self.ZARR_LOCATIONS[dataset],
+            storage_options={"requester_pays": True},
+        ).band_data
+        xarr.rio.write_crs("EPSG:4326", inplace=True)
+        xarr.name = dataset.get_field_name()
+
+        if geometry is not None:
+            return self._clip_xarr_to_geometry(xarr, geometry)
+        return xarr
+
+    def _clip_xarr_to_geometry(self, xarr, geom):
+        sliced = xarr.sel(
+            x=slice(geom.bounds[0], geom.bounds[2]),
+            y=slice(geom.bounds[3], geom.bounds[1]),
+        )
+        if "band" in sliced.dims:
+            sliced = sliced.squeeze("band")
+
+        geojson = mapping(geom)
+        clipped = sliced.rio.clip([geojson])
+        return clipped
